static_cast double -> size_t

+++ dynamic_programming/memoization/nth_fibonacci_number.cpp
@@ -42,7 +42,7 @@ constexpr auto nth_fibonacci = [](uint8_t n) -> size_t {
   constexpr double psi = -0.6'180'339'887'498'949'025;
   constexpr double sqrt_five = 2.23'606'797'749'979;
 
-  return (std::pow(phi, n) - std::pow(psi, n)) / sqrt_five;
+  return static_cast<size_t>((std::pow(phi, n) - std::pow(psi, n)) / sqrt_five);
 };
 
 TEST_CASE("nth_fibonacci_map") {
